Removed unused imports

+++ src/main/java/info/u_team/u_mod/gui/GuiPulverizer.java
@@ -1,8 +1,7 @@
 package info.u_team.u_mod.gui;
 
 import info.u_team.u_mod.UConstants;
-import info.u_team.u_mod.container.ContainerBase;
-import info.u_team.u_mod.container.ContainerPulverizer;
+import info.u_team.u_mod.container.*;
 import info.u_team.u_mod.resource.EnumModeTab;
 import info.u_team.u_mod.tilentity.pulverizer.TileEntityPulverizer;
 import info.u_team.u_team_core.container.UContainer;
@@ -22,10 +21,10 @@ public class GuiPulverizer extends UGuiContainer {
 	
 	@Override
 	protected void drawInBackground(EnumModeTab tab, int mouseX, int mouseY, int x_offset, int y_offset) {
-		if(tab == EnumModeTab.NORMAL)
+		if (tab == EnumModeTab.NORMAL)
 			this.drawTexturedModalRect(x_offset + 47, y_offset + 28, 0, 166, Math.round(64 * (100 - tile.getField(0)) / 100), 7);
-		if(tab == EnumModeTab.ENERGY)
-			this.drawTexturedModalRect(x_offset + 9, y_offset + 8, 0, 166, Math.round((float)158 * (float)((float)tile.impl_energy / (float)tile.getStorage().getMaxEnergyStored())), 20);
+		if (tab == EnumModeTab.ENERGY)
+			this.drawTexturedModalRect(x_offset + 9, y_offset + 8, 0, 166, Math.round(158 * ((float) tile.impl_energy / (float) tile.getStorage().getMaxEnergyStored())), 20);
 	}
 	
 }
